backend/common/requirements.txt
# requirements.txt

anyio==4.2.0
arrow==1.3.0
atlastk==0.13.3
attrs==23.2.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
exceptiongroup==1.2.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
jmespath==1.0.1
loguru==0.7.2
outcome==1.3.0.post0
playwright==1.41.2
pycparser==2.21
pyee==11.0.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
selenium==4.17.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tortoise==0.1.1
trio==0.24.0
trio-websocket==0.11.1
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
urllib3==2.2.0
win32-setctime==1.1.0
wsproto==1.2.0
